Moved Components to their respective folders
Moved the pop up menus to be in separate files from the CalendarPage.tsx such that they can be tested and reused in future pages that use events.
